refactor: divid Upload component to multiple components by type
